Make sure skbs that need fragmenting have dst info set
When tunnel security reprocessing is enabled, skb's arrive without
the skb_dst info set. This works ok for pkts that fit the mtu
determined in nmring_tg4(). However, if the packet is bigger it will
be sent to ip_do_fragment() and that assumes skb_dst is set and
de-references it, leading to a crash.

NAT being enabled seems to play a part - at least with the configuration
in the CR. I will do some more investigating to work out how this
impacts the problem.

Now, if the packet is going to be fragmented and skb_dst is not set
ipt_ipv4_route() is called to set it. A route lookup is costly so this
may not be the best solution?? However, I don't understand the code
well enough to suggest something else.

I'll de-AW+-afy this commit message once reviewed :)