Prevent crash when packet needs fragmenting and skb_dst not set
The ip_do_fragment() function depends on skb_dst info being set in
the skb. If this information isn't available drop the packet.

Note that icmp_send() also requires skb_dst but fails gracefully
if it is not set.

Additionally, removed use of skb_pagelen() from MTU calculation when
skb_dst is not set. This function gives a value relating to the size
of the current packet and seems to vary depending on how the packet
has been handled so is not approriate for MTU calculations. Instead
use the netmap buffer size allowing for headroom and the addition of
an ethernet header.