Fix generic netmap driver for the Octeon TX2
The ARX900S has a minimum of 4096 TX ring slots, which is larger than the
max which netmap allows. As the eth driver does not allow to reduce
this number. To mitigate this we now use the generic_ring_size parameter,
instead of the HW ring size parameter if the HW ring is too large.