Add queue-mapping to SKB
Socket buffers returned to the kernel can optionally have a queue-mapped
to them. This enables the kernel driver to TX them out of the core that
they arrived on, provided cores are mapped to queues 1 to 1.

Rx queues are recorded on SKBs with an increment, as a queue map of 0
means none was recorded.