mvpp2: Allow napi to finish
When enabling netmap processing while traffic is
running can result in the interruption of the
kernel NAPI process. This can result in the counts
for free rx descriptors and used rx descriptors to
be incorrect.

If this exception is seen, fall back to  NAPI. Netmap
process begins with a small initial RX disruption.