Rework deletion of ipt_netmap_priv
A crash can occur if rx/txsync is happening at the same time as
the rings are being released by deleting the iptables target.

The find_priv_from_na and find_ifc_pipe calls should be within locks
as they can access data that will be under contention. The returned
pointer needs some kind of protection. This is done with a reference
counter so that ipt_netmap_priv is deleted when either the sync ends
or the target is destroyed.