linux/configure: make driver tests depend on driver build
The new build system of Intel drivers generates kcompat files
that are then included in the driver's headers. If we run our
own driver-specific tests in parallel, we may end up reading
the kcompat files while they are being benerated. As a general
solution, we add a Makefile dependency between all the
$DRIVER-specific netmap tests and the build-$DRIVER target.
As a bonus, the netmap $DRIVER-specific tests will not run if $DRIVER
cannot be built.