Don't use AX25 pointer for netmap
Instead of using a pointer for a protocol we don't use (AX25), use the
ml_priv pointer in the net_device structure. This is intended as a
private pointer, so it matches what netmap wants. There is still a
danger that another protocol may want this pointer for its own private
data, but this is detected and reported if it happens.

Even though the pointer is now ml_priv instaed of ax25, the old #ifdef
is still used to make the patch as small as possible.